requirements.txt
Django~=3.1.4
django-allauth~=0.41.0
django-widget-tweaks~=1.4.8
Pillow